+++ lib/app/common/presentaion/widgets/article_card.dart
@@ -1,15 +1,15 @@
 import 'package:cached_network_image/cached_network_image.dart';
 import 'package:flutter/material.dart';
 import 'package:intl/intl.dart';
-import 'package:ziggle/app/core/theme/text.dart';
+import 'package:ziggle/app/common/presentaion/widgets/article_tags.dart';
+import 'package:ziggle/app/common/presentaion/widgets/button.dart';
+import 'package:ziggle/app/common/presentaion/widgets/d_day.dart';
+import 'package:ziggle/app/core/themes/text.dart';
 import 'package:ziggle/app/core/utils/functions/calculate_date_delta.dart';
-import 'package:ziggle/app/core/values/colors.dart';
+import 'package:ziggle/app/core/values/palette.dart';
 import 'package:ziggle/app/core/values/shadows.dart';
 import 'package:ziggle/app/data/enums/article_type.dart';
 import 'package:ziggle/app/data/model/article_summary_response.dart';
-import 'package:ziggle/app/global_widgets/article_tags.dart';
-import 'package:ziggle/app/global_widgets/button.dart';
-import 'package:ziggle/app/global_widgets/d_day.dart';
 import 'package:ziggle/gen/strings.g.dart';
 
 const kArticleCardHeight = 180.0;
