+++ lib/app/common/presentaion/widgets/bottom_sheet.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:ziggle/app/core/values/colors.dart';
+import 'package:ziggle/app/core/values/palette.dart';
 import 'package:ziggle/app/core/values/shadows.dart';
 
 class ZiggleBottomSheet extends StatelessWidget {
