+++ lib/app/modules/article/page.dart
@@ -3,10 +3,10 @@ import 'dart:math';
 import 'package:cached_network_image/cached_network_image.dart';
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
-import 'package:ziggle/app/core/theme/text.dart';
-import 'package:ziggle/app/core/values/colors.dart';
+import 'package:ziggle/app/common/presentaion/widgets/bottom_sheet.dart';
+import 'package:ziggle/app/core/themes/text.dart';
+import 'package:ziggle/app/core/values/palette.dart';
 import 'package:ziggle/app/core/values/shadows.dart';
-import 'package:ziggle/app/global_widgets/bottom_sheet.dart';
 import 'package:ziggle/app/modules/article/article_body.dart';
 import 'package:ziggle/app/modules/article/controller.dart';
 import 'package:ziggle/app/modules/article/page_spinner.dart';
