+++ lib/app/modules/auth/presentation/pages/splash_page.dart
@@ -1,0 +1,10 @@
+import 'package:flutter/material.dart';
+
+class SplashPage extends StatelessWidget {
+  const SplashPage({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const SizedBox.shrink();
+  }
+}
