+++ lib/app/modules/root/page.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
+import 'package:ziggle/app/common/presentaion/widgets/button.dart';
 import 'package:ziggle/app/core/routes/routes.dart';
-import 'package:ziggle/app/core/theme/text.dart';
-import 'package:ziggle/app/core/values/colors.dart';
-import 'package:ziggle/app/global_widgets/button.dart';
+import 'package:ziggle/app/core/themes/text.dart';
+import 'package:ziggle/app/core/values/palette.dart';
 import 'package:ziggle/app/modules/home/page.dart';
 import 'package:ziggle/app/modules/root/controller.dart';
 import 'package:ziggle/app/modules/search/page.dart';
