+++ lib/app/modules/write/page.dart
@@ -2,11 +2,11 @@ import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 import 'package:textfield_tags/textfield_tags.dart';
-import 'package:ziggle/app/core/theme/text.dart';
-import 'package:ziggle/app/core/values/colors.dart';
+import 'package:ziggle/app/common/presentaion/widgets/button.dart';
+import 'package:ziggle/app/common/presentaion/widgets/text_form_field.dart';
+import 'package:ziggle/app/core/themes/text.dart';
+import 'package:ziggle/app/core/values/palette.dart';
 import 'package:ziggle/app/data/enums/article_type.dart';
-import 'package:ziggle/app/global_widgets/button.dart';
-import 'package:ziggle/app/global_widgets/text_form_field.dart';
 import 'package:ziggle/app/modules/write/controller.dart';
 import 'package:ziggle/app/modules/write/images_picker.dart';
 import 'package:ziggle/gen/strings.g.dart';
